Abstract markdown parsing and add extensions
